refactor(retry): extract IsTransientFlapsError to flapsutil
Moves the transient-error classifier from strategy_bluegreen.go into a
shared internal/flapsutil/retry.go so upcoming resilience work on other
deploy paths (rolling update, blue teardown, health-check polls, etc.)
can reuse the same classification rules without duplicating the logic
per call site.

No behaviour change: strategy_bluegreen.go delegates to
flapsutil.IsTransientFlapsError instead of its private copy, and the
existing test coverage moves to internal/flapsutil/retry_test.go.
